+++ src/applications/claims-status/components/StemClaimListItem.jsx
@@ -1,22 +1,23 @@
-import PropTypes from 'prop-types';
 import React from 'react';
-import { Link } from 'react-router';
-import moment from 'moment';
-import recordEvent from 'platform/monitoring/record-event';
+import PropTypes from 'prop-types';
+
+import { recordEvent } from '@department-of-veterans-affairs/platform-monitoring/exports';
+
+import { buildDateFormatter } from '../utils/helpers';
+import ClaimCard from './ClaimCard';
+
+const formatDate = buildDateFormatter('MMMM d, yyyy');
 
 export default function StemClaimListItem({ claim }) {
   if (!claim.attributes.automatedDenial) {
     return null;
   }
-  const formattedDeniedAtDate = moment(claim.attributes.deniedAt).format(
-    'MMMM D, YYYY',
-  );
-  const formattedReceiptDate = moment(claim.attributes.submittedAt).format(
-    'MMMM D, YYYY',
-  );
+
+  const formattedDeniedAtDate = formatDate(claim.attributes.deniedAt);
+  const formattedReceiptDate = formatDate(claim.attributes.submittedAt);
 
   const handlers = {
-    openClaimClick: () =>
+    onClick: () => {
       recordEvent({
         event: 'cta-action-link-click',
         'action-link-type': 'secondary',
@@ -24,38 +25,30 @@ export default function StemClaimListItem({ claim }) {
         'action-link-icon-color': 'blue',
         'claim-type': 'STEM Scholarship',
         'claim-last-updated-date': formattedDeniedAtDate,
-        'claim-submitted-date': formattedDeniedAtDate,
+        'claim-submitted-date': formattedReceiptDate,
         'claim-status': 'Denied',
-      }),
+      });
+    },
   };
 
+  const ariaLabel = `View details for claim submitted on ${formattedReceiptDate}`;
+  const href = `your-stem-claims/${claim.id}/status`;
+
   return (
-    <div className="claim-list-item-container">
-      <h2 className="claim-list-item-header-v2 vads-u-font-size--h3">
-        Edith Nourse Rogers STEM Scholarship application
-        <br />
-        updated on {formattedDeniedAtDate}
-      </h2>
+    <ClaimCard
+      title="Edith Nourse Rogers STEM Scholarship application"
+      subtitle={`Received on ${formattedReceiptDate}`}
+    >
       <div className="card-status">
-        <div className="status-circle closed-claim" />
-        <p>
-          <strong>Status:</strong> Denied
-        </p>
+        <p>Status: Denied</p>
+        <p>Last updated on: {formattedDeniedAtDate}</p>
       </div>
-      <div className="card-status">
-        <p>
-          <strong>Received on:</strong> {formattedReceiptDate}
-        </p>
-      </div>
-      <Link
-        aria-label={`View details of claim received ${formattedReceiptDate}`}
-        className="vads-c-action-link--blue"
-        to={`your-stem-claims/${claim.id}/status`}
-        onClick={handlers.openClaimClick}
-      >
-        View details
-      </Link>
-    </div>
+      <ClaimCard.Link
+        ariaLabel={ariaLabel}
+        href={href}
+        onClick={handlers.onClick}
+      />
+    </ClaimCard>
   );
 }
