+++ src/applications/claims-status/tests/components/ClaimListItem.unit.spec.jsx
@@ -52,7 +52,7 @@ describe('<ClaimsListItem>', () => {
         decisionLetterSent: false,
         developmentLetterSent: true,
         documentsNeeded: false,
-        status: 'INITIAL_REVIEW',
+        status: 'EVIDENCE_GATHERING_REVIEW_DECISION',
       },
     };
 
@@ -60,33 +60,21 @@ describe('<ClaimsListItem>', () => {
     expect(screen.getByText('We sent you a development letter')).to.exist;
   });
 
-  it('should show development letter flag', () => {
-    const claim = {
-      id: 1,
-      attributes: {
-        decisionLetterSent: false,
-        developmentLetterSent: true,
-        documentsNeeded: false,
-        status: 'INITIAL_REVIEW',
-      },
-    };
-
-    const screen = render(<ClaimsListItem claim={claim} />);
-    expect(screen.getByText('We sent you a development letter')).to.exist;
-  });
-
-  it('should show decision letter flag', () => {
+  it('should show decision letter flag decisionLetterSent is true, but not render the other flags', () => {
     const claim = {
       id: 1,
       attributes: {
         decisionLetterSent: true,
-        developmentLetterSent: false,
-        documentsNeeded: false,
+        developmentLetterSent: true,
+        documentsNeeded: true,
         status: 'INITIAL_REVIEW',
       },
     };
 
     const screen = render(<ClaimsListItem claim={claim} />);
+    expect(screen.queryByText('We sent you a development letter')).to.be.null;
+    expect(screen.queryByText('An item in the claim needs your attention')).to
+      .be.null;
     expect(screen.getByText('You have a decision letter ready')).to.exist;
   });
 
@@ -102,7 +90,8 @@ describe('<ClaimsListItem>', () => {
     };
 
     const screen = render(<ClaimsListItem claim={claim} />);
-    expect(screen.getByText('Items need attention')).to.exist;
+    expect(screen.getByText('An item in the claim needs your attention')).to
+      .exist;
   });
 
   it('should not show any flags when closed', () => {
@@ -120,22 +109,6 @@ describe('<ClaimsListItem>', () => {
     expect(screen.queryByRole('listitem')).not.to.exist;
   });
 
-  it('should render a status circle with the `closed` class', () => {
-    const claim = {
-      id: 1,
-      attributes: {
-        status: 'COMPLETE',
-      },
-    };
-
-    const tree = shallow(<ClaimsListItem claim={claim} />);
-    const circle = tree.find('.status-circle').first();
-
-    expect(circle.hasClass('open-claim')).to.be.false;
-    expect(circle.hasClass('closed-claim')).to.be.true;
-    tree.unmount();
-  });
-
   it('should render a link to the claim status page', () => {
     const claim = {
       id: 1,
@@ -146,10 +119,10 @@ describe('<ClaimsListItem>', () => {
 
     const tree = shallow(<ClaimsListItem claim={claim} />);
     const linkProps = tree
-      .find('Link')
+      .find('ClaimCardLink')
       .first()
       .props();
-    expect(linkProps.to).to.equal(`your-claims/${claim.id}/status`);
+    expect(linkProps.href).to.equal(`your-claims/${claim.id}/status`);
     tree.unmount();
   });
 
@@ -166,8 +139,7 @@ describe('<ClaimsListItem>', () => {
     };
 
     const screen = render(<ClaimsListItem claim={claim} />);
-    expect(screen.getByText(/updated on August 20, 2019/i)).to.exist;
-    expect(screen.getByText(/Received on/i)).to.exist;
-    expect(screen.getByText(/February 10, 2019/i)).to.exist;
+    expect(screen.getByText(/Last updated: August 20, 2019/i)).to.exist;
+    expect(screen.getByText(/Received on February 10, 2019/i)).to.exist;
   });
 });
