+++ src/applications/claims-status/tests/components/StemClaimListItem.unit.spec.jsx
@@ -16,7 +16,7 @@ describe('<StemClaimListItem>', () => {
 
   it('should render a denied STEM claim', () => {
     const tree = shallow(<StemClaimListItem claim={defaultClaim} />);
-    expect(tree.find('h2').length).to.equal(1);
+    expect(tree.find('ClaimCard').length).to.equal(1);
     tree.unmount();
   });
 
@@ -34,33 +34,25 @@ describe('<StemClaimListItem>', () => {
   });
 
   it('should render updated on and submitted on with proper date', () => {
-    const tree = shallow(<StemClaimListItem claim={defaultClaim} />);
-    expect(tree.find('h2').text()).to.contain('updated on March 2, 2021');
+    const wrapper = shallow(<StemClaimListItem claim={defaultClaim} />);
     expect(
-      tree
-        .find('p')
-        .last()
+      wrapper
+        .find('ClaimCard')
+        .shallow()
+        .find('.submitted-on')
         .text(),
-    ).to.equal('Received on: March 1, 2021');
-    tree.unmount();
+    ).to.equal('Received on March 1, 2021');
+    wrapper.unmount();
   });
 
   it('should show denied status', () => {
     const tree = shallow(<StemClaimListItem claim={defaultClaim} />);
     expect(
       tree
-        .find('.status-circle + p')
+        .find('.card-status p')
         .first()
         .text(),
     ).to.equal('Status: Denied');
-    tree.unmount();
-  });
-
-  it('should render a status circle with the `closed` class', () => {
-    const tree = shallow(<StemClaimListItem claim={defaultClaim} />);
-    const circle = tree.find('.status-circle').first();
-    expect(circle.hasClass('open-claim')).to.be.false;
-    expect(circle.hasClass('closed-claim')).to.be.true;
     tree.unmount();
   });
 });
